+++ RobotSource/include/Drivetrain.h
@@ -13,9 +13,9 @@ class Drivetrain : private SerialLink {
     void setSpeed(int) const;
     void move(Direction) const;
     void move(Direction, int) const;
-    void nudge(Direction, float);
-    void travel(Direction, int, float, bool);
-    void travel(float, float, float, float);
+    void nudge(Direction, double);
+    void travel(Direction, int, double, bool);
+    void travel(double, double, double, double);
 };
 
 #endif//DRIVETRAIN_H
