+++ RobotSource/src/Coord.cpp
@@ -2,11 +2,11 @@
 
 #include <cmath>
 
-Coord::Coord(float x, float y) {
+Coord::Coord(double x, double y) {
   this->x = x;
   this->y = y;
 }
 
-float Coord::dist(Coord c) const {
+double Coord::dist(Coord c) const {
   return std::sqrt(std::pow((x-c.x),2)+std::pow((y-c.y),2));
 }
